# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for For Democracy",
        "title": "For Democracy API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Support Team",
            "email": "support@for-democracy"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/missions": {
            "post": {
                "description": "Create a mission",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "missions"
                ],
                "summary": "Create a mission",
                "parameters": [
                    {
                        "description": "Mission object that needs to be created",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validators.Mission"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/validators.Mission"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/missions/{id}": {
            "put": {
                "description": "Update a mission",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "missions"
                ],
                "summary": "Update a mission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Mission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Mission properties that needs to be updated",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validators.Mission"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/validators.Mission"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a mission",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "missions"
                ],
                "summary": "Delete a mission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Mission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/missions/{id}/join": {
            "post": {
                "description": "Join a mission",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "missions"
                ],
                "summary": "Join a mission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Mission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Mission properties that needs to be updated",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validators.UserMission"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.GroupUserMission"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/missions/{id}/leave": {
            "post": {
                "description": "Leave a mission",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "missions"
                ],
                "summary": "Leave a mission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Mission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "enum.Difficulty": {
            "type": "string",
            "enum": [
                "trivial",
                "easy",
                "medium",
                "challenging",
                "hard",
                "extreme",
                "suicide_mission",
                "impossible",
                "helldive"
            ],
            "x-enum-varnames": [
                "Trivial",
                "Easy",
                "Medium",
                "Challenging",
                "Hard",
                "Extreme",
                "SuicideMission",
                "Impossible",
                "Helldive"
            ]
        },
...
        "enum.ObjectiveType": {
            "type": "string",
            "enum": [
                "terminate_illegal_broadcast",
                "pump_fuel_to_icbm",
                "upload_escape_pod_data",
                "conduct_geological_survey",
                "launch_icbm",
                "retrieve_valuable_data",
                "emergency_evacuation",
                "spread_democracy",
                "eliminate_brood_commanders",
                "purge_hatcheries",
                "activate_e710_pumps",
                "nuke_nursery",
                "activate_terminid_control_system",
                "blitz_search_and_destroy_terminids",
                "eliminate_chargers",
                "eradicate_terminid_swarm",
                "eliminate_bile_titans",
                "enable_e710_extraction",
                "eliminate_devastators",
                "eliminate_automaton_hulks",
                "sabotage_supply_bases",
                "destroy_transmission_network",
                "eradicate_automaton_forces",
                "blitz_search_and_destroy_automatons",
                "sabotage_air_base",
                "eliminate_automaton_factory_strider",
                "destroy_command_bunkers"
            ],
            "x-enum-varnames": [
                "TerminateIllegalBroadcast",
                "PumpFuelToICBM",
                "UploadEscapePodData",
                "ConductGeologicalSurvey",
                "LaunchICBM",
                "RetrieveValuableData",
                "EmergencyEvacuation",
                "SpreadDemocracy",
                "EliminateBroodCommanders",
                "PurgeHatcheries",
                "ActivateE710Pumps",
                "NukeNursery",
                "ActivateTerminidControlSystem",
                "BlitzSearchAndDestroyTerminids",
                "EliminateChargers",
                "EradicateTerminidSwarm",
                "EliminateBileTitans",
                "EnableE710Extraction",
                "EliminateDevastators",
                "EliminateAutomatonHulks",
                "SabotageSupplyBases",
                "DestroyTransmissionNetwork",
                "EradicateAutomatonForces",
                "BlitzSearchAndDestroyAutomatons",
                "SabotageAirBase",
                "EliminateAutomatonFactoryStrider",
                "DestroyCommandBunkers"
            ]
        },
        "enum.Role": {
            "type": "string",
            "enum": [
                "admin",
                "user"
            ],
            "x-enum-varnames": [
                "Admin",
                "User"
            ]
        },
        "enum.StratagemKeys": {
            "type": "string",
            "enum": [
                "up",
                "right",
                "down",
                "left"
            ],
            "x-enum-varnames": [
                "Up",
                "Right",
                "Down",
                "Left"
            ]
        },
        "enum.StratagemType": {
            "type": "string",
            "enum": [
                "supply",
                "mission",
                "defensive",
                "offensive"
            ],
            "x-enum-varnames": [
                "Supply",
                "Mission",
                "Defensive",
                "Offensive"
            ]
        },
        "enum.StratagemUseType": {
            "type": "string",
            "enum": [
                "self",
                "team",
                "shared"
            ],
            "x-enum-varnames": [
                "Self",
                "Team",
                "Shared"
            ]
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "model.Group": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "difficulty": {
                    "$ref": "#/definitions/enum.Difficulty"
                },
                "groupUsers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GroupUser"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "missions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Mission"
                    }
                },
                "name": {
                    "type": "string"
                },
                "planet": {
                    "$ref": "#/definitions/model.Planet"
                },
                "planetID": {
                    "type": "integer"
                },
                "public": {
                    "type": "boolean"
                },
                "startAt": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.GroupUser": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "group": {
                    "$ref": "#/definitions/model.Group"
                },
                "groupID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "owner": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "model.GroupUserMission": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "groupUser": {
                    "$ref": "#/definitions/model.GroupUser"
                },
                "groupUserID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "mission": {
                    "$ref": "#/definitions/model.Mission"
                },
                "missionID": {
                    "type": "integer"
                },
                "stratagems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Stratagem"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "model.Mission": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "group": {
                    "$ref": "#/definitions/model.Group"
                },
                "groupID": {
                    "type": "integer"
                },
                "groupUserMissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GroupUserMission"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "instructions": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "objectiveTypes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/enum.ObjectiveType"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "model.Stratagem": {
            "type": "object",
            "properties": {
                "activation": {
                    "type": "integer"
                },
                "codeName": {
                    "type": "string"
                },
                "cooldown": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "groupUserMissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GroupUserMission"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "imageURL": {
                    "type": "string"
                },
                "keys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/enum.StratagemKeys"
                    }
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/enum.StratagemType"
                },
                "updatedAt": {
                    "type": "string"
                },
                "useCount": {
                    "type": "integer"
                },
                "useType": {
                    "$ref": "#/definitions/enum.StratagemUseType"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "groupUsers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GroupUser"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/enum.Role"
                },
                "steamId": {
                    "type": "string"
                },
                "tokens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Token"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "server.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "validators.Mission": {
            "type": "object",
            "required": [
                "name",
                "objectiveTypes"
            ],
            "properties": {
                "groupID": {
                    "type": "integer"
                },
                "instructions": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "objectiveTypes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/enum.ObjectiveType"
                    }
                }
            }
        },
        "validators.UserMission": {
            "type": "object",
            "required": [
                "stratagems"
            ],
            "properties": {
                "stratagems": {
                    "type": "array",
                    "maxItems": 4,
                    "items": {
                        "type": "integer"
                    }
                }
            }
        }
    }
}